# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Novus Markets Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── Users ──────────────────────────────────────────────────────────────────
...
  /accounts/summary:
    get:
      operationId: getAccountSummary
      tags: [accounts]
      summary: Get account summary (balance, equity, margin, free margin)
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSummary"

  # ── Markets ─────────────────────────────────────────────────────────────────
...
  /admin/settings:
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin platform settings (fees, spread)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: kycStatus
          in: query
          schema:
            type: string
            enum: [unverified, pending, verified]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/balance:
    patch:
      operationId: adjustUserBalance
      tags: [admin]
      summary: Manually adjust user balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceAdjustment"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /admin/users/{id}/kyc:
    patch:
      operationId: updateKycDecision
      tags: [admin]
      summary: Approve or reject KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycDecision"
      responses:
        "200":
          description: KYC decision saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /admin/deposits:
    get:
      operationId: getAdminDeposits
      tags: [admin]
      summary: Get all pending deposits (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /admin/deposits/{id}/approve:
    post:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/deposits/{id}/reject:
    post:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        preferredCurrency:
          type: string
          default: USD
        role:
          type: string
          enum: [user, admin]
        twoFaEnabled:
          type: boolean
        notificationsEnabled:
          type: boolean
        theme:
          type: string
          enum: [dark, light]
          default: dark
        createdAt:
          type: string
          format: date-time
...
    KycStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [unverified, pending, verified, rejected]
        documentType:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
...
    Account:
      type: object
      required: [id, userId, type, balance, currency, leverage, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [real, demo]
        balance:
          type: number
        currency:
          type: string
        leverage:
          type: integer
        createdAt:
          type: string
          format: date-time
    AccountSummary:
      type: object
      required: [balance, equity, margin, freeMargin, marginLevel, openPnl]
      properties:
        balance:
          type: number
        equity:
          type: number
        margin:
          type: number
        freeMargin:
          type: number
        marginLevel:
          type: number
        openPnl:
          type: number

    # ── Markets ────────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, userId, type, amount, fee, netAmount, method, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        method:
          type: string
          enum: [mpesa, crypto_usdt, crypto_btc, card, bank_transfer]
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        reference:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
...
    PlatformSettings:
      type: object
      required: [depositFeeRate, spreadMarkup, minDeposit, maxLeverage, demoBalance]
      properties:
        depositFeeRate:
          type: number
          description: Deposit fee percentage (e.g. 0.02 = 2%)
        spreadMarkup:
          type: number
          description: Additional spread markup in pips
        minDeposit:
          type: number
        maxLeverage:
          type: integer
        demoBalance:
          type: number
        cryptoWalletUsdt:
          type: string
        cryptoWalletBtc:
          type: string
        mpesaPaybill:
          type: string

    # ── Admin ──────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, email, firstName, lastName, role, kycStatus, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [user, admin]
        kycStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        totalDeposits:
          type: number
        totalTrades:
          type: integer
        realBalance:
          type: number
        createdAt:
          type: string
          format: date-time
    AdminSettingsUpdate:
      type: object
      properties:
        depositFeeRate:
          type: number
        spreadMarkup:
          type: number
        minDeposit:
          type: number
        maxLeverage:
          type: integer
        cryptoWalletUsdt:
          type: string
        cryptoWalletBtc:
          type: string
        mpesaPaybill:
          type: string
    BalanceAdjustment:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: number
        reason:
          type: string
    KycDecision:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [verified, rejected]
        rejectionReason:
          type: string