# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Novus Markets Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  depositFeeRate?: number;
  spreadMarkup?: number;
  minDeposit?: number;
  maxLeverage?: number;
  cryptoWalletUsdt?: string;
  cryptoWalletBtc?: string;
  mpesaPaybill?: string;
}