# lib/api-zod/src/generated/types/kycDecision.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Novus Markets Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { KycDecisionStatus } from './kycDecisionStatus';

export interface KycDecision {
  status: KycDecisionStatus;
  rejectionReason?: string;
}