# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Novus Markets Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserTheme } from './userTheme';

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
  preferredCurrency?: string;
  role: UserRole;
  twoFaEnabled?: boolean;
  notificationsEnabled?: boolean;
  theme?: UserTheme;
  createdAt: Date;
}